requirements.txt
openai
PyPDF2==3.0.0
python-dotenv==1.0.0
streamlit==1.30.0